// src/app/pages/corralons/components/corralons-table/corralons.interface.ts
export interface CorralonsInterface {
   idcorralon?: number;
   fecha?: string;
   hora?: string;
   fechaSalida?: string;
   horaSalida?: string;
   infraccionNumero?: number;
   corralonNombre?: string;
   motivo?: string;
   estado_idestado?: number;
   permisotaxiasignado_idpermisotaxiasignado?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}